// libs/store/Discount/actions-type.tsx
export enum DiscountActionType {
    SEARCH_DISCOUNT = 'discount/SEARCH_DISCOUNT',
    SEARCH_DISCOUNT_SUCCESS = 'discount/SEARCH_DISCOUNT_SUCCESS',
    SEARCH_DISCOUNT_FAILED = 'discount/SEARCH_DISCOUNT_FAILED',
    ADD_DISCOUNT = 'discount/ADD_DISCOUNT',
    ADD_DISCOUNT_SUCCESS = 'discount/ADD_DISCOUNT_SUCCESS',
    ADD_DISCOUNT_FAILED = 'discount/ADD_DISCOUNT_FAILED',
    GET_DISCOUNTS = 'discount/GET_DISCOUNTS',
    GET_DISCOUNTS_SUCCESS = 'discount/GET_DISCOUNTS_SUCCESS',
    GET_DISCOUNTS_FAILED = 'discount/GET_DISCOUNTS_FAILED'
}